route-service-working

+++ server/route-service/src/main/java/com/codeflowdb/route/DjkstraAlgorithm.java
@@ -6,9 +6,9 @@ public class DjkstraAlgorithm {
 
     public static class RouteEdge {
         private String destination;
-        private int cost;
+        private double cost;
 
-        public RouteEdge(String destination, int cost) {
+        public RouteEdge(String destination, double cost) {
             this.destination = destination;
             this.cost = cost;
         }
@@ -23,22 +23,22 @@ public class DjkstraAlgorithm {
     }
 
     public static List<String> findShortestPath(String start, String end, Map<String, List<RouteEdge>> graph) {
-        Map<String, Integer> distance = new HashMap<>();
+        Map<String, Double> distance = new HashMap<>();
         Map<String, String> prev = new HashMap<>();
-        PriorityQueue<String> pq = new PriorityQueue<>(Comparator.comparingInt(distance::get));
+        PriorityQueue<String> pq = new PriorityQueue<>(Comparator.comparingDouble(distance::get));
 
         for (String node : graph.keySet()) {
-            distance.put(node, Integer.MAX_VALUE);
+            distance.put(node, Double.MAX_VALUE);
             prev.put(node, null);
         }
 
-        distance.put(start, 0);
+        distance.put(start, 0.0);
         pq.add(start);
 
         while (!pq.isEmpty()) {
             String current = pq.poll();
             for (RouteEdge edge : graph.getOrDefault(current, new ArrayList<>())) {
-                int newDist = distance.get(current) + edge.cost;
+                double newDist = distance.get(current) + edge.cost;
                 if (newDist < distance.get(edge.destination)) {
                     distance.put(edge.destination, newDist);
                     prev.put(edge.destination, current);
